+++ src/components/Usuario.js
@@ -1,0 +1,9 @@
+
+
+export function Usuario (){
+    return (
+        <div>
+            <p>Lucas Calabria</p>
+        </div>
+    )
+}
