+++ src/styles.css
@@ -72,3 +72,8 @@ footer {
 h4 {
   margin-left: 20px;
 }
+
+#bighead {
+  height: 0.5rem;
+  width: 0.5rem;
+}
